Update JS query instrumentation to instrument promise interface
Starting in node-postgres
[5.1.0](https://github.com/brianc/node-postgres/blob/master/CHANGELOG.md#v510),
the javascript interface has allowed for promise-style calling of the
query method of the Client object, where no callback is passed and a
promise which resolves with the query's result is returned from the
Client.query call.

However, since the new relic instrumentation has not been aware of the
promise interface, if you use the promise interface with the current
version of the new relic agent, queries show up in new relic as taking
0ms.  If I understand the datastore shim API correctly, this update
should record the time between when Client.query is called and when the
promise resolves as the amount of time that a query took, when the
promise-based interface is used.

The native client instrumentation was left alone because it still only
allows callback-style use.